refactor: separate webserver code into modules
